+++ pyresample/resampler.py
@@ -20,12 +20,13 @@ from __future__ import annotations
 
 import logging
 import os
-from abc import ABC, abstractmethod
 from functools import lru_cache
 from numbers import Number
 from typing import Union
 
 import numpy as np
+
+from .slicer import Slicer, _enumerate_chunk_slices
 
 try:
     import dask.array as da
@@ -38,15 +39,11 @@ try:
 except ImportError:
     xr = None
 
-from pyproj.transformer import Transformer
-
 from pyresample.geometry import (
     AreaDefinition,
     CoordinateDefinition,
     IncompatibleAreas,
-    InvalidArea,
     SwathDefinition,
-    get_geostationary_bounding_box_in_proj_coords,
 )
 
 from .future.resamplers.resampler import hash_dict
@@ -259,197 +256,3 @@ def _enumerate_dst_area_chunks(dst_area, chunks):
                       "chunk-shape": chunk_shape,
                       }
         yield block_info, target_geo_def
-
-
-def _enumerate_chunk_slices(chunks):
-    """Enumerate chunks with slices."""
-    for position in np.ndindex(tuple(map(len, (chunks)))):
-        slices = []
-        for pos, chunk in zip(position, chunks):
-            chunk_size = chunk[pos]
-            offset = sum(chunk[:pos])
-            slices.append(slice(offset, offset + chunk_size))
-
-        yield (position, slices)
-
-
-def bil2(src_area, dst_area, data, indices_xy, block_info=None):
-    """Bilinear interpolation implementation for resample_blocks."""
-    del src_area, dst_area
-    x_indices, y_indices = indices_xy
-    if block_info:
-        y_slice, x_slice = block_info[0]["array-location"][-2:]
-        x_indices -= x_slice.start
-        y_indices -= y_slice.start
-    mask = np.isnan(y_indices)
-    x_indices = np.nan_to_num(x_indices, 0)
-    y_indices = np.nan_to_num(y_indices, 0)
-
-    w_l, l_a = np.modf(y_indices.clip(0, data.shape[-2] - 1))
-    w_p, p_a = np.modf(x_indices.clip(0, data.shape[-1] - 1))
-
-    l_a = l_a.astype(int)
-    p_a = p_a.astype(int)
-    l_b = np.clip(l_a + 1, 1, data.shape[-2] - 1)
-    p_b = np.clip(p_a + 1, 1, data.shape[-1] - 1)
-
-    res = ((1 - w_l) * (1 - w_p) * data[..., l_a, p_a] +
-           (1 - w_l) * w_p * data[..., l_a, p_b] +
-           w_l * (1 - w_p) * data[..., l_b, p_a] +
-           w_l * w_p * data[..., l_b, p_b])
-    res = np.where(mask, np.nan, res)
-    return res
-
-
-class Slicer(ABC):
-    """Abstract Slicer and Slicer factory class, returning a AreaSlicer or a SwathSlicer based on the first area type.
-
-    Provided an Area-to-crop and an Area-to-contain, a Slicer provides methods
-    to find slices that enclose Area-to-contain inside Area-to-crop.
-    """
-
-    def __new__(cls, area_to_crop, area_to_contain):
-        """Create a Slicer for cropping *area_to_crop* based on *area_to_contain*."""
-        if cls is Slicer:
-            if isinstance(area_to_crop, SwathDefinition):
-                return SwathSlicer(area_to_crop, area_to_contain)
-            elif isinstance(area_to_crop, AreaDefinition):
-                return AreaSlicer(area_to_crop, area_to_contain)
-            else:
-                raise NotImplementedError("Don't know how to slice a " + str(type(area_to_crop)))
-        else:
-            return super().__new__(cls)
-
-    def __init__(self, area_to_crop, area_to_contain):
-        """Set up the Slicer."""
-        self.area_to_crop = area_to_crop
-        self.area_to_contain = area_to_contain
-        self._transformer = Transformer.from_crs(self.area_to_contain.crs, self.area_to_crop.crs)
-
-    def get_slices(self):
-        """Get the slices to crop *area_to_crop* enclosing *area_to_contain*."""
-        poly = self.get_polygon()
-        return self.get_slices_from_polygon(poly)
-
-    @abstractmethod
-    def get_polygon(self):
-        """Get the shapely Polygon corresponding to *area_to_contain*."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_slices_from_polygon(self, poly):
-        """Get the slices based on the polygon."""
-        raise NotImplementedError
-
-
-class SwathSlicer(Slicer):
-    """A Slicer for cropping SwathDefinitions."""
-
-    def get_polygon(self):
-        """Get the shapely Polygon corresponding to *area_to_contain* in lon/lat coordinates."""
-        from shapely.geometry import Polygon
-        x, y = self.area_to_contain.get_bbox_coords(10)
-        poly = Polygon(zip(*self._transformer.transform(x, y)))
-        return poly
-
-    def get_slices_from_polygon(self, poly):
-        """Get the slices based on the polygon."""
-        intersecting_chunk_slices = []
-        for smaller_poly, slices in _get_chunk_polygons_for_area_to_crop(self.area_to_crop):
-            if smaller_poly.intersects(poly):
-                intersecting_chunk_slices.append(slices)
-        if not intersecting_chunk_slices:
-            raise IncompatibleAreas
-        return self._assemble_slices(intersecting_chunk_slices)
-
-    @staticmethod
-    def _assemble_slices(chunk_slices):
-        """Assemble slices to one slice per dimension."""
-        lines, cols = zip(*chunk_slices)
-        line_slice = slice(min(slc.start for slc in lines), max(slc.stop for slc in lines))
-        col_slice = slice(min(slc.start for slc in cols), max(slc.stop for slc in cols))
-        slices = col_slice, line_slice
-        return slices
-
-
-@lru_cache(maxsize=10)
-def _get_chunk_polygons_for_area_to_crop(area_to_crop):
-    """Get the polygons for each chunk of the area_to_crop."""
-    res = []
-    from shapely.geometry import Polygon
-    chunks = np.array(area_to_crop.lons.data.chunksize) // 2
-    src_chunks = da.core.normalize_chunks(chunks, area_to_crop.shape)
-    for _position, (line_slice, col_slice) in _enumerate_chunk_slices(src_chunks):
-        line_slice = expand_slice(line_slice)
-        col_slice = expand_slice(col_slice)
-        smaller_swath = area_to_crop[line_slice, col_slice]
-        lons, lats = smaller_swath.get_edge_lonlats(10)
-        lons = np.hstack(lons)
-        lats = np.hstack(lats)
-        smaller_poly = Polygon(zip(lons, lats))
-        res.append((smaller_poly, (line_slice, col_slice)))
-    return res
-
-
-def expand_slice(small_slice):
-    """Expand slice by one."""
-    return slice(max(small_slice.start - 1, 0), small_slice.stop + 1, small_slice.step)
-
-
-class AreaSlicer(Slicer):
-    """A Slicer for cropping AreaDefinitions."""
-
-    def get_polygon(self):
-        """Get the shapely Polygon corresponding to *area_to_contain* in projection coordinates."""
-        from shapely.geometry import Polygon
-        x, y = self.area_to_contain.get_bbox_coords(frequency=10)
-        # before_poly = Polygon(zip(x, y)).buffer(np.max(target_area.resolution))
-        # x, y = zip(*before_poly.exterior.coords)
-        if self.area_to_crop.is_geostationary:
-            x_geos, y_geos = get_geostationary_bounding_box_in_proj_coords(self.area_to_crop, 360)
-            x_geos, y_geos = self._transformer.transform(x_geos, y_geos, direction='INVERSE')
-            geos_poly = Polygon(zip(x_geos, y_geos))
-            poly = Polygon(zip(x, y))
-            poly = poly.intersection(geos_poly)
-            if poly.is_empty:
-                raise IncompatibleAreas('No slice on area.')
-            x, y = zip(*poly.exterior.coords)
-        poly = Polygon(zip(*self._transformer.transform(x, y)))
-        return poly
-
-    def get_slices_from_polygon(self, poly):
-        """Get the slices based on the polygon."""
-        # We take a little margin around the polygon to ensure all needed pixels will be included.
-        try:
-            bounds = poly.buffer(np.max(self.area_to_contain.resolution)).bounds
-        except ValueError as err:
-            raise InvalidArea(str(err))
-        bounds = self._sanitize_polygon_bounds(bounds)
-        slice_x, slice_y = self._create_slices_from_bounds(bounds)
-        return slice_x, slice_y
-
-    def _sanitize_polygon_bounds(self, bounds):
-        """Reset the bounds within the shape of the area."""
-        try:
-            (minx, miny, maxx, maxy) = bounds
-        except ValueError:
-            raise IncompatibleAreas('No slice on area.')
-        x_bounds, y_bounds = self.area_to_crop.get_array_coordinates_from_projection_coordinates(np.array([minx, maxx]),
-                                                                                                 np.array([miny, maxy]))
-        y_size, x_size = self.area_to_crop.shape
-        if np.all(x_bounds < 0) or np.all(y_bounds < 0) or np.all(x_bounds >= x_size) or np.all(y_bounds >= y_size):
-            raise IncompatibleAreas('No slice on area.')
-        return x_bounds, y_bounds
-
-    @staticmethod
-    def _create_slices_from_bounds(bounds):
-        """Create slices from bounds."""
-        x_bounds, y_bounds = bounds
-        try:
-            slice_x = slice(int(np.floor(max(np.min(x_bounds), 0))),
-                            int(np.ceil(np.max(x_bounds))))
-            slice_y = slice(int(np.floor(max(np.min(y_bounds), 0))),
-                            int(np.ceil(np.max(y_bounds))))
-        except OverflowError:
-            raise IncompatibleAreas("Area not within finite bounds.")
-        return expand_slice(slice_x), expand_slice(slice_y)
