+++ pyresample/test/test_resample_blocks.py
@@ -21,200 +21,10 @@
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Tests for the DaskResampler."""
 
-import unittest
 import dask.array as da
-import xarray as xr
 import numpy as np
-from pyresample.resampler import Slicer
-from pyresample.geometry import AreaDefinition, SwathDefinition, IncompatibleAreas, InvalidArea
-import pytest
 
-
-class TestAreaSlicer(unittest.TestCase):
-    """Test the get_slice method for AreaSlicers."""
-
-    def setUp(self):
-        """Set up the test case."""
-        self.dst_area = AreaDefinition('euro40', 'euro40', None,
-                                       {'proj': 'stere', 'lon_0': 14.0,
-                                        'lat_0': 90.0, 'lat_ts': 60.0,
-                                        'ellps': 'bessel'},
-                                       102, 102,
-                                       (-2717181.7304994687, -5571048.14031214,
-                                        1378818.2695005313, -1475048.1403121399))
-
-    def test_source_area_covers_dest_area(self):
-        """Test source area covers dest area."""
-        src_area = AreaDefinition('dst', 'dst area', None,
-                                  {'ellps': 'WGS84', 'h': '35785831', 'proj': 'geos'},
-                                  100, 100,
-                                  (5550000.0, 5550000.0, -5550000.0, -5550000.0))
-        slicer = Slicer(src_area, self.dst_area)
-        x_slice, y_slice = slicer.get_slices()
-        assert x_slice.start > 0 and x_slice.stop <= 100
-        assert y_slice.start > 0 and y_slice.stop <= 100
-
-    def test_source_area_does_not_cover_dest_area_entirely(self):
-        """Test source area does not cover dest area entirely."""
-        src_area = AreaDefinition('dst', 'dst area', None,
-                                  {'ellps': 'WGS84', 'h': '35785831', 'proj': 'geos'},
-                                  100, 100,
-                                  (5550000.0, 4440000.0, -5550000.0, -6660000.0))
-
-        slicer = Slicer(src_area, self.dst_area)
-        x_slice, y_slice = slicer.get_slices()
-        assert x_slice.start > 0 and x_slice.stop < 100
-        assert y_slice.start > 0 and y_slice.stop >= 100
-
-    def test_source_area_does_not_cover_dest_area_at_all(self):
-        """Test source area does not cover dest area at all."""
-        src_area = AreaDefinition('dst', 'dst area', None,
-                                  {'ellps': 'WGS84', 'h': '35785831', 'proj': 'geos'},
-                                  80, 100,
-                                  (5550000.0, 3330000.0, -5550000.0, -5550000.0))
-
-        slicer = Slicer(src_area, self.dst_area)
-        with pytest.raises(IncompatibleAreas):
-            slicer.get_slices()
-
-    # def test_dest_area_is_outside_source_area_domain(self):
-    #     """Test dest area is outside the source area domain (nan coordinates)."""
-    #     src_area = AreaDefinition('dst', 'dst area', None,
-    #                               {'ellps': 'WGS84', 'h': '35785831', 'proj': 'geos'},
-    #                               100, 100,
-    #                               (5550000.0, 5550000.0, -5550000.0, -5550000.0))
-    #     dst_area = AreaDefinition('merc', 'merc', None,
-    #                               {'proj': 'merc', 'lon_0': 120.0,
-    #                                'lat_0': 0,
-    #                                'ellps': 'bessel'},
-    #                               102, 102,
-    #                               (-100000, -100000,
-    #                                100000, 100000))
-    #     slicer = Slicer(src_area, dst_area)
-    #     with pytest.raises(IncompatibleAreas):
-    #         slicer.get_slices()
-
-    def test_barely_touching_chunks_intersection(self):
-        """Test that barely touching chunks generate slices on intersection."""
-        src_area = AreaDefinition('dst', 'dst area', None,
-                                  {'ellps': 'WGS84', 'h': '35785831', 'proj': 'geos'},
-                                  100, 100,
-                                  (5550000.0, 5550000.0, -5550000.0, -5550000.0))
-        dst_area = AreaDefinition('moll', 'moll', None,
-                                  {
-                                      'ellps': 'WGS84',
-                                      'lon_0': '0',
-                                      'proj': 'moll',
-                                      'units': 'm'
-                                  },
-                                  102, 102,
-                                  (-18040095.6961, 4369712.0686,
-                                   18040095.6961, 9020047.8481))
-        slicer = Slicer(src_area, dst_area)
-        x_slice, y_slice = slicer.get_slices()
-        assert x_slice.start > 0 and x_slice.stop < 100
-        assert y_slice.start > 0 and y_slice.stop >= 100
-
-    def test_slicing_an_area_with_infinite_bounds(self):
-        """Test slicing an area with infinite bounds."""
-        src_area = AreaDefinition('dst', 'dst area', None,
-                                  {'ellps': 'WGS84', 'proj': 'merc'},
-                                  100, 100,
-                                  (-10000.0, -10000.0, 0.0, 0.0))
-
-        dst_area = AreaDefinition('moll', 'moll', None,
-                                  {
-                                      'ellps': 'WGS84',
-                                      'lon_0': '0',
-                                      'proj': 'moll',
-                                      'units': 'm'
-                                  },
-                                  102, 102,
-                                  (-100000.0, -4369712.0686,
-                                   18040096.0, 9020047.8481))
-
-        slicer = Slicer(src_area, dst_area)
-        with pytest.raises(InvalidArea):
-            slicer.get_slices()
-
-
-class TestSwathSlicer(unittest.TestCase):
-    """Test the get_slice function when input is a swath."""
-
-    def setUp(self):
-        """Set up the test case."""
-        chunks = 10
-        self.dst_area = AreaDefinition('euro40', 'euro40', None,
-                                       {'proj': 'stere', 'lon_0': 14.0,
-                                        'lat_0': 90.0, 'lat_ts': 60.0,
-                                        'ellps': 'bessel'},
-                                       102, 102,
-                                       (-2717181.7304994687, -5571048.14031214,
-                                        1378818.2695005313, -1475048.1403121399))
-        self.src_area = AreaDefinition(
-            'omerc_otf',
-            'On-the-fly omerc area',
-            None,
-            {'alpha': '8.99811271718795',
-             'ellps': 'sphere',
-             'gamma': '0',
-             'k': '1',
-             'lat_0': '0',
-             'lonc': '13.8096029486222',
-             'proj': 'omerc',
-             'units': 'm'},
-            50, 100,
-            (-1461111.3603, 3440088.0459, 1534864.0322, 9598335.0457)
-        )
-
-        lons, lats = self.src_area.get_lonlats(chunks=chunks)
-        lons = xr.DataArray(lons.persist())
-        lats = xr.DataArray(lats.persist())
-        self.src_swath = SwathDefinition(lons, lats)
-
-    def test_slicer_init(self):
-        """Test slicer initialization."""
-        slicer = Slicer(self.src_area, self.dst_area)
-        assert slicer.area_to_crop == self.src_area
-        assert slicer.area_to_contain == self.dst_area
-
-    def test_source_swath_slicing_does_not_return_full_dataset(self):
-        """Test source area covers dest area."""
-        slicer = Slicer(self.src_swath, self.dst_area)
-        x_slice, y_slice = slicer.get_slices()
-        assert (x_slice, y_slice) == (slice(0, 36), slice(14, 91))
-        assert x_slice.start == 0
-        assert x_slice.stop == 36
-        assert y_slice.start == 14
-        assert y_slice.stop == 91
-
-    def test_source_area_slicing_does_not_return_full_dataset(self):
-        """Test source area covers dest area."""
-        slicer = Slicer(self.src_area, self.dst_area)
-        x_slice, y_slice = slicer.get_slices()
-        assert x_slice.start == 0
-        assert x_slice.stop == 35
-        assert y_slice.start == 16
-        assert y_slice.stop == 94
-
-    def test_area_get_polygon_returns_a_polygon(self):
-        """Test getting a polygon returns a polygon."""
-        from shapely.geometry import Polygon
-        slicer = Slicer(self.src_area, self.dst_area)
-        poly = slicer.get_polygon()
-        assert isinstance(poly, Polygon)
-
-    def test_swath_get_polygon_returns_a_polygon(self):
-        """Test getting a polygon returns a polygon."""
-        from shapely.geometry import Polygon
-        slicer = Slicer(self.src_swath, self.dst_area)
-        poly = slicer.get_polygon()
-        assert isinstance(poly, Polygon)
-
-    def test_cannot_slice_a_string(self):
-        """Test that we cannot slice a string."""
-        with pytest.raises(NotImplementedError):
-            Slicer("my_funky_area", self.dst_area)
+from pyresample.geometry import AreaDefinition
 
 
 class TestResampleBlocksArea2Area:
@@ -389,8 +199,8 @@ class TestResampleBlocksArea2Area:
 
     def test_resample_blocks_can_generate_gradient_indices(self):
         """Test resample blocks can generate gradient indices."""
-        from pyresample.resampler import resample_blocks
         from pyresample.gradient import gradient_resampler_indices
+        from pyresample.resampler import resample_blocks
 
         chunks = 40
 
@@ -403,8 +213,13 @@ class TestResampleBlocksArea2Area:
 
     def test_resample_blocks_can_gradient_resample(self):
         """Test resample_blocks can do gradient resampling."""
+        from pyresample.gradient import (
+            block_bilinear_interpolator,
+            gradient_resampler,
+            gradient_resampler_indices,
+        )
         from pyresample.resampler import resample_blocks
-        from pyresample.gradient import gradient_resampler_indices, gradient_resampler
+
         chunksize = 40
 
         some_array = da.arange(self.src_area.shape[0] * self.src_area.shape[1]).astype(float)
@@ -413,9 +228,9 @@ class TestResampleBlocksArea2Area:
         indices = resample_blocks(self.src_area, self.dst_area, gradient_resampler_indices,
                                   chunks=(2, chunksize, chunksize), dtype=float)
         np.testing.assert_allclose(gradient_resampler_indices(self.src_area, self.dst_area), indices)
-        from pyresample.resampler import bil2
-        res = resample_blocks(self.src_area, self.dst_area, bil2, some_array, dst_arrays=[indices],
-                              chunks=(chunksize, chunksize), dtype=some_array.dtype)
+
+        res = resample_blocks(self.src_area, self.dst_area, block_bilinear_interpolator, some_array,
+                              dst_arrays=[indices], chunks=(chunksize, chunksize), dtype=some_array.dtype)
         np.testing.assert_allclose(gradient_resampler(some_array.compute(), self.src_area, self.dst_area), res)
 
     def test_resample_blocks_passes_kwargs(self):
@@ -463,7 +278,7 @@ class TestResampleBlocksArea2Area:
 
         dst_array = da.arange(np.product(self.dst_area.shape)).reshape(self.dst_area.shape).rechunk(40)
         res = resample_blocks(self.src_area, self.dst_area, fun, dst_arrays=[dst_array], chunks=40, dtype=float)
-        res = res.compute()
+        _ = res.compute()
 
     def test_resample_blocks_can_pass_block_info_about_target(self):
         """Test resample_blocks can pass block_info about the target chunk."""
@@ -484,7 +299,7 @@ class TestResampleBlocksArea2Area:
 
         dst_array = da.arange(np.product(self.dst_area.shape)).reshape(self.dst_area.shape).rechunk(40)
         res = resample_blocks(self.src_area, self.dst_area, fun, dst_arrays=[dst_array], chunks=40, dtype=float)
-        res = res.compute()
+        _ = res.compute()
 
     def test_resample_blocks_supports_3d_dst_arrays(self):
         """Test resample_blocks supports 3d dst_arrays."""
@@ -501,6 +316,7 @@ class TestResampleBlocksArea2Area:
         np.testing.assert_allclose(res[:, 40:], dst_array[0, :, 40:])
 
     # test_multiple_inputs
-    # test output type
     # test 3d resampling
     # test chunks on non-xy dimensions
+    # test passing own fill value
+    # test "auto" chunk size
